+++ CHANGELOG.md
@@ -12,13 +12,14 @@ Changes are categorized as follows:
 
 ## [Unreleased]
 
-* 5 minor enhancements:
+* 6 minor enhancements:
 
   * docs: add github paths
   * deps: add debug
   * config: set json spaces 2
   * middleware: nested parsers
   * middleware: improve error handler
+  * deps: add root-model
 
 ## 2018-05-03, Version 1.0.4